added get user example endpoint

+++ messaging-project-server/src/models.rs
@@ -1,8 +1,9 @@
 use serde::{Deserialize, Serialize};
+use sqlx::FromRow;
 
-#[derive(Debug, Deserialize, Serialize, Clone)]
+#[derive(Debug, Deserialize, Serialize, Clone, FromRow)]
 pub struct User {
-  pub user_id: u64,
+  pub user_id: i32,
   pub username: String,
   pub password: String
 }
